requirements.txt
cryptography
Flask==2.2.2
numpy==1.24.0
Pillow==9.3.0
PyGithub==1.57
waitress==2.1.2
